requirements.txt
# pymongo==4.1.1
mysql-connector-python==8.0.29
streamlit
numpy